Modification and implementation for compatibility with PSR-4 standard
