+++ src/Resources/Subscriptions.php
@@ -1,0 +1,93 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: anvargear
+ * Date: 19/03/17
+ * Time: 12:49 PM
+ */
+
+namespace Epayco\Resources;
+
+
+use Epayco\Resource;
+
+class Subscriptions extends Resource
+{
+
+    /**
+     * Create subscription
+     * @param  object $options data client and plan
+     * @return object
+     */
+    public function create($options = null)
+    {
+        return $this->request(
+            "POST",
+            "/recurring/v1/subscription/create",
+            $api_key = $this->epayco->api_key,
+            $options,
+            $private_key = $this->epayco->private_key,
+            $test = $this->epayco->test,
+            $switch = false,
+            $lang = $this->epayco->lang
+        );
+    }
+
+    /**
+     * Get subscription from id
+     * @param  String $uid id subscription
+     * @return object
+     */
+    public function get($uid)
+    {
+        return $this->request(
+            "GET",
+            "/recurring/v1/subscription/" . $uid . "/" . $this->epayco->api_key  . "/",
+            $api_key = $this->epayco->api_key,
+            $options = null,
+            $private_key = $this->epayco->private_key,
+            $test = $this->epayco->test,
+            $switch = false,
+            $lang = $this->epayco->lang
+        );
+    }
+
+    /**
+     * Get all subscriptions from client epayco
+     * @return object
+     */
+    public function getList()
+    {
+        return $this->request(
+            "GET",
+            "/recurring/v1/subscriptions/" . $this->epayco->api_key,
+            $api_key = $this->epayco->api_key,
+            $options = null,
+            $private_key = $this->epayco->private_key,
+            $test = $this->epayco->test,
+            $switch = false,
+            $lang = $this->epayco->lang
+        );
+    }
+
+    /**
+     * Cancel active subscription
+     * @param  String $uid id subscription
+     * @return object
+     */
+    public function cancel($uid)
+    {
+        return $this->request(
+            "POST",
+            "/recurring/v1/subscription/cancel",
+            $api_key = $this->epayco->api_key,
+            $options = array(
+                "id" => $uid
+            ),
+            $private_key = $this->epayco->private_key,
+            $test = $this->epayco->test,
+            $switch = false,
+            $lang = $this->epayco->lang
+        );
+    }
+}
